requirements.txt
streamlit>=1.30.0
google-generativeai>=0.4.0
pandas>=2.2.0
numpy>=1.26.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.17.0
openpyxl>=3.1.2
xlrd>=2.0.1
python-dotenv>=1.0.0
chardet>=5.2.0
scikit-learn>=1.3.0

# LangChain Ecosystem - Core Framework
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# LangGraph - State Machine & Advanced Workflows
langgraph>=0.2.0

# LangSmith - Observability & Monitoring
langsmith>=0.1.0

# LangChain Integrations
langchain-google-genai>=2.0.0
langchain-openai>=0.2.0

# Additional LangChain Components
langchain-experimental>=0.3.0
pydantic>=2.5.0